pyproject.toml
[tool.poetry]
name = "geotl"
version = "0.1.0"
description = ""
authors = ["ChethanUK <chethanuk@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
geopandas = "^0.12.2"
psycopg2-binary = "^2.9.5"
openpyxl = "^3.0.10"
h3pandas = "^0.2.3"
duckdb = "^0.6.1"
uszipcode = "^1.0.1"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.12.0"
isort = "^5.11.4"
requests = "^2.28.1"
psycopg2-binary = "^2.9.5"
keplergl = "^0.3.2"
jupyterlab = "^3.5.2"
geemap = "^0.19.6"
gnssmapper = "^0.0.1-alpha.6"
unlzw3 = "^0.2.2"
pyogrio = "^0.5.1"
geopy = "^2.3.0"
osmnx = "^1.3.0"
pyrosm = "^0.6.1"
r5py = "^0.0.4"
kaleido = "0.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
